+++ package.json
@@ -40,7 +40,8 @@
   "devDependencies": {
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.0.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/react-hooks": "^7.0.2",
     "@types/jest-axe": "^3.5.9",
     "@types/js-cookie": "^3.0.6",
     "@types/lodash": "^4.17.16",
@@ -53,6 +54,7 @@
     "jest-axe": "^9.0.0",
     "lint-staged": "^15.2.11",
     "tailwindcss": "^3.4.1",
+    "ts-node": "^10.9.2",
     "typescript": "^5"
   }
 }
