+++ test/mocks/mocked-posts.ts
@@ -1,0 +1,262 @@
+export const MOCKED_POSTS = [
+  {
+    post: {
+      uri: 'at://did:plc:7jotvgmlhdanzanutftrixhl/app.bsky.feed.post/3lbssmv24tk22',
+      cid: 'bafyreieaf222n24a56r3rlfnxnrvwrdfsuyzytzbk4lm5vv5so7xgllete',
+      author: {
+        did: 'did:plc:7jotvgmlhdanzanutftrixhl',
+        handle: 'crackedlemonade.bsky.social',
+        displayName: ' Cracked Lemonade',
+        avatar:
+          'https://cdn.bsky.app/img/avatar/plain/did:plc:7jotvgmlhdanzanutftrixhl/bafkreihxp4dnsk6wbyimqjqn4s5twxzyjoihxfvezubajt5bpeo46uv4mm@jpeg',
+        associated: {
+          chat: {
+            allowIncoming: 'following',
+          },
+        },
+        labels: [
+          {
+            src: 'did:plc:7jotvgmlhdanzanutftrixhl',
+            uri: 'at://did:plc:7jotvgmlhdanzanutftrixhl/app.bsky.actor.profile/self',
+            cid: 'bafyreigmiyakr7sdsvvzzhmsrez72irdk4z5o3wzo4qtshrrednd5bc2ee',
+            val: '!no-unauthenticated',
+            cts: '1970-01-01T00:00:00.000Z',
+          },
+        ],
+        createdAt: '2023-04-25T16:27:16.609Z',
+      },
+      record: {
+        $type: 'app.bsky.feed.post',
+        createdAt: '2024-11-26T00:44:09.567Z',
+        embed: {
+          $type: 'app.bsky.embed.record',
+          record: {
+            cid: 'bafyreicywiodzzyohyjzxjk72bezrnyr5bzglhevau4n72ncwkpj6noyze',
+            uri: 'at://did:plc:c6k2ozisr2bbitsafwwyistf/app.bsky.feed.post/3lbssgbkif22r',
+          },
+        },
+        langs: ['en'],
+        text: 'Kendrick unintentionally fed us so much dropping an amazing album with no promotion and this residual Drake melt down',
+      },
+      embed: {
+        $type: 'app.bsky.embed.record#view',
+        record: {
+          $type: 'app.bsky.embed.record#viewRecord',
+          uri: 'at://did:plc:c6k2ozisr2bbitsafwwyistf/app.bsky.feed.post/3lbssgbkif22r',
+          cid: 'bafyreicywiodzzyohyjzxjk72bezrnyr5bzglhevau4n72ncwkpj6noyze',
+          author: {
+            did: 'did:plc:c6k2ozisr2bbitsafwwyistf',
+            handle: 'ruemcclammyhand.bsky.social',
+            displayName: 'regular meghan 나영지 🍉',
+            avatar:
+              'https://cdn.bsky.app/img/avatar/plain/did:plc:c6k2ozisr2bbitsafwwyistf/bafkreie352i773dgdhy3gs6rk4ey433cfmg5wlmuicw7dfykwfqv4yk3uy@jpeg',
+            associated: {
+              chat: {
+                allowIncoming: 'following',
+              },
+            },
+            labels: [
+              {
+                src: 'did:plc:c6k2ozisr2bbitsafwwyistf',
+                uri: 'at://did:plc:c6k2ozisr2bbitsafwwyistf/app.bsky.actor.profile/self',
+                cid: 'bafyreidfvgmx2z5pwm6awj5rb4tnpmxzn4rotf5u5ewyyrto6plnz4gnim',
+                val: '!no-unauthenticated',
+                cts: '1970-01-01T00:00:00.000Z',
+              },
+            ],
+            createdAt: '2023-04-23T19:23:52.606Z',
+          },
+          value: {
+            $type: 'app.bsky.feed.post',
+            createdAt: '2024-11-26T00:40:27.805Z',
+            langs: ['en'],
+            text: 'feels illegal to watch someone (drake) melt down like this',
+          },
+          labels: [],
+          likeCount: 42,
+          replyCount: 3,
+          repostCount: 2,
+          quoteCount: 1,
+          indexedAt: '2024-11-26T00:40:28.056Z',
+          embeds: [],
+        },
+      },
+      replyCount: 0,
+      repostCount: 0,
+      likeCount: 4,
+      quoteCount: 0,
+      indexedAt: '2024-11-26T00:44:10.553Z',
+      labels: [],
+    },
+  },
+  {
+    post: {
+      uri: 'at://did:plc:ztxuq2ef5bm2r45oachkr63r/app.bsky.feed.post/3lbssmteqtc2c',
+      cid: 'bafyreia2txuurd3zjxi66lv52dizji2fihinnooyw2bpf66b7yjnhmwqzy',
+      author: {
+        did: 'did:plc:ztxuq2ef5bm2r45oachkr63r',
+        handle: 'maybeantonio.bsky.social',
+        displayName: 'Antonio',
+        avatar:
+          'https://cdn.bsky.app/img/avatar/plain/did:plc:ztxuq2ef5bm2r45oachkr63r/bafkreifex5iqph5ps67voyhb7g63rdfwg6eeut5jgyk7whu6rs2k33n6yi@jpeg',
+        labels: [],
+        createdAt: '2024-11-08T21:07:48.685Z',
+      },
+      record: {
+        $type: 'app.bsky.feed.post',
+        createdAt: '2024-11-26T00:44:07.817Z',
+        langs: ['en'],
+        text: "Drake got to realize he will never be a the Artistic and creative level that Kendrick is n that's ok. Know ya place.",
+      },
+      replyCount: 0,
+      repostCount: 1,
+      likeCount: 7,
+      quoteCount: 1,
+      indexedAt: '2024-11-26T00:44:07.264Z',
+      labels: [],
+    },
+  },
+  {
+    post: {
+      uri: 'at://did:plc:ofegrjclspaablwuhtdxjqv6/app.bsky.feed.post/3lbssm7zzys2s',
+      cid: 'bafyreihb2dqrnvw6berjv4iffed5wvem3ftkn6hmh6zs2yqfjilyjxflbe',
+      author: {
+        did: 'did:plc:ofegrjclspaablwuhtdxjqv6',
+        handle: 'coolclarks.bsky.social',
+        displayName: '',
+        avatar:
+          'https://cdn.bsky.app/img/avatar/plain/did:plc:ofegrjclspaablwuhtdxjqv6/bafkreig6o5vyshxknrr3vl7vrbhqkcckwyysyyieuugz7xfqauyyp6wxfy@jpeg',
+        labels: [],
+        createdAt: '2024-11-21T02:18:57.469Z',
+      },
+      record: {
+        $type: 'app.bsky.feed.post',
+        createdAt: '2024-11-26T00:43:47.544Z',
+        langs: ['en'],
+        text: 'Dude started his own “Streamgate” alleging Universal and Spotify were colluding to help Kendrick? Tsk tsk',
+      },
+      replyCount: 0,
+      repostCount: 0,
+      likeCount: 0,
+      quoteCount: 0,
+      indexedAt: '2024-11-26T00:43:48.550Z',
+      labels: [],
+    },
+  },
+  {
+    post: {
+      uri: 'at://did:plc:jvecmt4ic6dmcndlnsk64fiw/app.bsky.feed.post/3lbsslroqqc2z',
+      cid: 'bafyreifjwihwdninalxeejoluzncdg3pnemz6dwwhh6dfq6bfthii6ey2q',
+      author: {
+        did: 'did:plc:jvecmt4ic6dmcndlnsk64fiw',
+        handle: 'jehnuhsuhs.bsky.social',
+        displayName: 'Genesis',
+        avatar:
+          'https://cdn.bsky.app/img/avatar/plain/did:plc:jvecmt4ic6dmcndlnsk64fiw/bafkreibyvmkxdezfdqauuj2v545q3qaouy3ze22rp2nz5pmpoxmh2khlsq@jpeg',
+        labels: [
+          {
+            src: 'did:plc:jvecmt4ic6dmcndlnsk64fiw',
+            uri: 'at://did:plc:jvecmt4ic6dmcndlnsk64fiw/app.bsky.actor.profile/self',
+            cid: 'bafyreibn7fwk552yibeispc6t5gnc6o5mmusz3ifjicn7mj3wiwxh7yyke',
+            val: '!no-unauthenticated',
+            cts: '2024-11-14T00:12:02.168Z',
+          },
+        ],
+        createdAt: '2024-11-14T00:12:04.293Z',
+      },
+      record: {
+        $type: 'app.bsky.feed.post',
+        createdAt: '2024-11-26T00:43:32.494Z',
+        embed: {
+          $type: 'app.bsky.embed.video',
+          aspectRatio: {
+            height: 720,
+            width: 720,
+          },
+          video: {
+            $type: 'blob',
+            ref: {
+              $link:
+                'bafkreigykn32bicuybcn4wmwb4hyrmqxfgvugfarod6jkguj5edvlr6vsi',
+            },
+            mimeType: 'video/mp4',
+            size: 987132,
+          },
+        },
+        facets: [
+          {
+            features: [
+              {
+                $type: 'app.bsky.richtext.facet#tag',
+                tag: 'megantheestallion',
+              },
+            ],
+            index: {
+              byteEnd: 316,
+              byteStart: 298,
+            },
+          },
+          {
+            features: [
+              {
+                $type: 'app.bsky.richtext.facet#tag',
+                tag: 'kendrick',
+              },
+            ],
+            index: {
+              byteEnd: 326,
+              byteStart: 317,
+            },
+          },
+        ],
+        langs: ['en'],
+        text: '“These niggas hate on BBLs and be walking ‘round with the same scars.” 🤝 “Let your core audience stomach that, then tell em where you get your abs from.”\n\n“Cosplay: Gangstas, Fake Ass: Accents ” 🤝 “You know I got language barriers (huh) it’s no accent you can sell me…”\n#megantheestallion #kendrick',
+      },
+      embed: {
+        $type: 'app.bsky.embed.video#view',
+        cid: 'bafkreigykn32bicuybcn4wmwb4hyrmqxfgvugfarod6jkguj5edvlr6vsi',
+        playlist:
+          'https://video.bsky.app/watch/did%3Aplc%3Ajvecmt4ic6dmcndlnsk64fiw/bafkreigykn32bicuybcn4wmwb4hyrmqxfgvugfarod6jkguj5edvlr6vsi/playlist.m3u8',
+        thumbnail:
+          'https://video.bsky.app/watch/did%3Aplc%3Ajvecmt4ic6dmcndlnsk64fiw/bafkreigykn32bicuybcn4wmwb4hyrmqxfgvugfarod6jkguj5edvlr6vsi/thumbnail.jpg',
+        aspectRatio: {
+          height: 720,
+          width: 720,
+        },
+      },
+      replyCount: 0,
+      repostCount: 7,
+      likeCount: 6,
+      quoteCount: 0,
+      indexedAt: '2024-11-26T00:43:33.656Z',
+      labels: [],
+    },
+  },
+  {
+    post: {
+      uri: 'at://did:plc:mlfxuy6chhdifek2xxdq72ro/app.bsky.feed.post/3lbsslpbg4s2f',
+      cid: 'bafyreihcoiahsjnx5jrmrktb7pweokrjpswyjwhafonz26x2z4zlxjqsnm',
+      author: {
+        did: 'did:plc:mlfxuy6chhdifek2xxdq72ro',
+        handle: 'coupdemain.bsky.social',
+        displayName: 'coup de main',
+        avatar:
+          'https://cdn.bsky.app/img/avatar/plain/did:plc:mlfxuy6chhdifek2xxdq72ro/bafkreihoa3nnsrmlp6wqcxof6b6hpazci4iz5z7jrprsel3vnxernq4ffy@jpeg',
+        labels: [],
+        createdAt: '2023-12-04T09:57:02.869Z',
+      },
+      record: {
+        $type: 'app.bsky.feed.post',
+        createdAt: '2024-11-26T00:43:29.960Z',
+        langs: ['en'],
+        text: 'the whiplash of going from the wicked soundtrack to new kendrick album',
+      },
+      replyCount: 0,
+      repostCount: 0,
+      likeCount: 2,
+      quoteCount: 0,
+      indexedAt: '2024-11-26T00:43:30.953Z',
+      labels: [],
+    },
+  },
+];
