+++ test/setupTests.ts
@@ -3,25 +3,29 @@ import { render as rtlRender } from '@testing-library/react';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import { Providers } from '@/contexts';
 import { ReactNode, AwaitedReactNode, JSX } from 'react';
+import { QueryClient } from '@tanstack/react-query';
+import {
+  renderHook as rtlRenderHook,
+  act as rtlAct,
+} from '@testing-library/react';
 
 expect.extend(toHaveNoViolations);
 
 let originalWarn: typeof console.warn;
-let originalDebug: typeof console.debug;
 
 beforeAll(() => {
-  // Stop warns and debugs from cluttering the test output
+  // Suppress warnings and debug messages during tests.
   originalWarn = console.warn;
-  originalDebug = console.debug;
   console.warn = jest.fn();
-  console.debug = jest.fn();
 });
 
 afterAll(() => {
   console.warn = originalWarn;
-  console.debug = originalDebug;
 });
 
+/**
+ * Custom render function that wraps the UI in our Providers.
+ */
 const render = (
   ui:
     | string
@@ -34,11 +38,22 @@ const render = (
     | null
     | undefined,
   { ...renderOptions } = {}
-) => {
-  return rtlRender(ui, { wrapper: Providers, ...renderOptions });
+) => rtlRender(ui, { wrapper: Providers, ...renderOptions });
+
+/**
+ * createQueryClient - Returns a new QueryClient with caching disabled for each test.
+ */
+const createQueryClient = () => new QueryClient();
+
+// Export our testing utilities.
+export {
+  render,
+  axe,
+  toHaveNoViolations,
+  rtlRenderHook as renderHook,
+  rtlAct as act,
+  createQueryClient,
 };
 
-// re-export everything
+// Re-export everything else from @testing-library/react.
 export * from '@testing-library/react';
-// override render method
-export { render, axe, toHaveNoViolations };
